Improved individual pet integration
