+++ src/components/UpcomingAppointments.js
@@ -50,9 +50,25 @@ async function createCalendar() {
   }
 }
 
+async function getMonthsEvents() {
+  const currentCalendar = [await AsyncStorage.getItem('@calendarID')];
+
+  const startDate = new Date();
+  const month = startDate.getMonth() + 1;
+  const endDate = new Date();
+  endDate.setMonth(month);
+
+  const monthsEvents = await
+  Calendar.getEventsAsync(currentCalendar, new Date(startDate), new Date(endDate));
+
+  return monthsEvents;
+}
+
 export default function UpcomingAppointments() {
   const dispatch = useDispatch();
   const defaultCalendar = useSelector((state) => state.calendar.calendarID);
+  const [monthsEvents, setMonthsEvents] = useState([]);
+  const [allEventsCollected, setEventsCollected] = useState(false);
 
   useEffect(() => {
     (async () => {
@@ -63,6 +79,14 @@ export default function UpcomingAppointments() {
       }
     })();
   }, []);
+
+  if (!allEventsCollected) {
+    (async () => {
+      const events = await getMonthsEvents();
+      setMonthsEvents(events);
+      setEventsCollected(true);
+    })();
+  }
 
   function addHour(date) {
     const endHour = new Date(date);
@@ -78,6 +102,8 @@ export default function UpcomingAppointments() {
     };
 
     await Calendar.createEventAsync(ID, newEvent);
+
+    setEventsCollected(false);
   }
 
   const [eventAdded, showEventAdded] = useState(false);
@@ -127,23 +153,18 @@ export default function UpcomingAppointments() {
         style={styles.healthDivider}
       />
       <View style={{ flex: 2 }}>
-        <View style={styles.appointmentPiece}>
-          <Text style={styles.appointmentText}>
-            Appointment Name
-          </Text>
-          <Text style={styles.appointmentText}>
-            Date
-          </Text>
-        </View>
 
-        <View style={styles.appointmentPiece}>
-          <Text style={styles.appointmentText}>
-            Appointment Name
-          </Text>
-          <Text style={styles.appointmentText}>
-            Date
-          </Text>
-        </View>
+        { monthsEvents.map((event, index) => (
+          // eslint-disable-next-line react/no-array-index-key
+          <View style={styles.appointmentPiece} key={`appointment-${index}`}>
+            <Text style={styles.appointmentText}>
+              {event.title}
+            </Text>
+            <Text style={styles.appointmentText}>
+              {dateFormat(event.startDate, 'm/dd @ h:MM tt')}
+            </Text>
+          </View>
+        )) }
 
       </View>
