+++ src/constants/DarkStyles.js
@@ -246,7 +246,7 @@ export default StyleSheet.create({
     paddingRight: 20,
   },
   usernameField: {
-    marginBottom: 50,
+    marginBottom: 20,
     justifyContent: 'center',
     backgroundColor: pawGreen,
   },
@@ -356,7 +356,7 @@ export default StyleSheet.create({
     borderColor: pawYellow,
     borderRadius: 80,
     overflow: 'hidden',
-    marginBottom: 40,
+    marginBottom: 20,
   },
   accountSwitch: {
     flexBasis: 'auto',
@@ -427,7 +427,7 @@ export default StyleSheet.create({
     backgroundColor: pawYellow,
     position: 'absolute',
     right: 120,
-    bottom: 40,
+    bottom: 20,
     padding: 8,
     paddingLeft: 9,
     borderRadius: 23,
@@ -468,6 +468,32 @@ export default StyleSheet.create({
     color: pawYellow,
     flexBasis: 'auto',
     paddingRight: 10,
+  },
+  medicalbutton: {
+    alignSelf: 'center',
+    width: (Dimensions.get('window').width - 20),
+    backgroundColor: pawLightGrey,
+    borderRadius: 25,
+    overflow: 'hidden',
+    marginBottom: 20,
+    height: 60,
+    flexDirection: 'row',
+    justifyContent: 'center',
+    alignItems: 'center',
+    paddingLeft: 18,
+    paddingRight: 20,
+  },
+  medicalText: {
+    fontSize: 24,
+    width: (Dimensions.get('window').width - 120),
+    fontWeight: '600',
+    textTransform: 'uppercase',
+    letterSpacing: 1,
+    fontFamily: 'QuicksandBold',
+    color: pawYellow,
+    flexBasis: 'auto',
+    textDecorationLine: 'underline',
+    textDecorationColor: pawGreen,
   },
   alertText: {
     fontSize: 24,
