+++ src/constants/Styles.js
@@ -255,7 +255,7 @@ export default StyleSheet.create({
     paddingRight: 20,
   },
   usernameField: {
-    marginBottom: 50,
+    marginBottom: 15,
     justifyContent: 'center',
     backgroundColor: pawPink,
   },
@@ -337,7 +337,30 @@ export default StyleSheet.create({
     alignItems: 'center',
     paddingLeft: 18,
     paddingRight: 20,
-
+  },
+  medicalbutton: {
+    alignSelf: 'center',
+    width: (Dimensions.get('window').width - 20),
+    backgroundColor: pawWhite,
+    borderRadius: 25,
+    overflow: 'hidden',
+    marginTop: 20,
+    marginBottom: 20,
+    justifyContent: 'center',
+    alignItems: 'center',
+    padding: 15,
+  },
+  medicalText: {
+    fontSize: 24,
+    width: (Dimensions.get('window').width - 120),
+    fontWeight: '600',
+    textTransform: 'uppercase',
+    letterSpacing: 1,
+    fontFamily: 'QuicksandBold',
+    color: pawGrey,
+    flexBasis: 'auto',
+    textDecorationLine: 'underline',
+    textDecorationColor: pawPink,
   },
   submittext: {
     alignSelf: 'center',
@@ -365,7 +388,7 @@ export default StyleSheet.create({
     borderColor: pawWhite,
     borderRadius: 80,
     overflow: 'hidden',
-    marginBottom: 40,
+    marginBottom: 20,
   },
   accountSwitch: {
     flexBasis: 'auto',
@@ -437,7 +460,7 @@ export default StyleSheet.create({
     backgroundColor: pawWhite,
     position: 'absolute',
     right: 120,
-    bottom: 40,
+    bottom: 20,
     padding: 8,
     paddingLeft: 9,
     borderRadius: 23,
